requirements.txt
# ansible==2.5
alembic==0.9.9
amqp==2.2.2
attrs==17.4.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
eventlet==0.22.1
flake8==3.5.0
greenlet==0.4.13
idna==2.6
kombu==4.1.0
ldap3==2.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
msgpack-python==0.5.6
mysqlclient==1.3.12
pluggy==0.6.0
py==1.5.3
pyasn1==0.4.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
tornado==4.5.3
urllib3==1.22
vine==1.1.4
yapf==0.21.0
PyMySQL==0.7.11
ansible==2.5.0
flower==0.9.2
